+++ desktop/packages/mullvad-vpn/test/e2e/route-object-models/user-interface-settings/selectors.ts
@@ -10,8 +10,20 @@ export const createSelectors = (page: Page) => ({
       hasText: label,
     }),
   autoStartSwitch: () => page.getByRole('switch', { name: 'Notifications' }),
-  monochromaticTrayIconSwitch: () => page.getByRole('switch', { name: 'Monochromatic tray icon' }),
-  unpinnedWindowSwitch: () => page.getByRole('switch', { name: 'Unpin app from taskbar' }),
+  monochromaticTrayIconSwitch: () => {
+    if (process.platform === 'darwin') {
+      return page.getByRole('switch', { name: 'Monochromatic menu bar icon' });
+    }
+
+    return page.getByRole('switch', { name: 'Monochromatic tray icon' });
+  },
+  unpinnedWindowSwitch: () => {
+    if (process.platform === 'win32') {
+      return page.getByRole('switch', { name: 'Unpin app from taskbar' });
+    }
+
+    return page.getByRole('switch', { name: 'Unpin app from menu bar' });
+  },
   startMinimizedSwitch: () => page.getByRole('switch', { name: 'Start minimized' }),
   animateMapSwitch: () => page.getByRole('switch', { name: 'Animate map' }),
 });